requirements.txt
flask~=3.1.0
flask_restful
flask_cors
openai~=1.61.0
pandas~=2.2.3

Flask-Cors~=5.0.0
python-dotenv~=1.0.1
python-docx~=1.1.2
